Fix legitimate compiler warning setting unsigned to negative.

+++ src/engine/JackDriver.h
@@ -72,7 +72,7 @@ private:
                 bool            unregister;
         };
 
-        volatile size_t         m_running;
+        volatile long int       m_running;
         jack_client_t*          m_jack_client{};
         QList<PortChannelPair*> m_inputs;
         QList<PortChannelPair*> m_outputs;
